Set global mocha timeout on prod and move utils
