+++ test/contracts/utils/index.js
@@ -1,5 +1,5 @@
 const { detectNetworkName } = require('./detectNetwork');
-const { connectContract, connectContracts } = require('./getContract');
+const { connectContract, connectContracts } = require('./connectContract');
 
 module.exports = {
 	detectNetworkName,
